Always send a final info line before reporting the best move
Also includes a couple of bits of reporting refactoring:

* PrincipalVariation must be sent back to search.rs in order to be
  reported the final time. In the future if we have thread voting, we
  may also want to report the PV of a thread which had the best move we
  send, if it is not the thread's best move.

* Added SearchInfo::from_ctx to clean up initialisation now it is being
  used in multiple places.

* Moved some fields into SearchStats so they can be initialised in
  from_ctx.

* Keep track of our completed depth so that it can be reported in the
  final info line.

Elo   | 1.67 +- 5.48 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.95 (-2.94, 2.94) [-10.00, 0.00]
Games | N: 4376 W: 1039 L: 1018 D: 2319
Penta | [27, 500, 1113, 521, 27]

+++ engine/src/engine/uci/responses.rs
@@ -43,7 +43,7 @@ pub struct InfoFields {
     pub(super) pv: Option<Vec<UciMove>>,
     pub(super) score: Option<InfoScore>,
     pub(super) wdl: Option<WdlProbabilities>,
-    pub(super) hashfull: Option<usize>,
+    pub(super) hashfull: Option<u64>,
     pub(super) nps: Option<u64>,
     pub(super) tbhits: Option<u64>,
     pub(super) string: Option<String>,
